# openapi.yaml
openapi: 3.1.0
info:
  title: Ad Performance Tracker API
  version: 2.0.0
  description: |
    Ad Performance Tracker — ระบบติดตามประสิทธิภาพโฆษณา
    รองรับ multi-page, multi-product per entry, HOT sales, CRM fields
  contact:
    name: Backend Oracle
servers:
  - url: https://ad-tracker-kappa.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /api/auth/test-token or NextAuth session
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        cost: { type: number }
        price: { type: number }
        stock: { type: integer }
      required: [id, name, cost, price, stock]
    Account:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
      required: [id, name]
    Page:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
      required: [id, name]
    EntryProduct:
      type: object
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        product:
          $ref: '#/components/schemas/Product'
      required: [productId, quantity]
    CalculatedFields:
      type: object
      description: Business metrics คำนวณจาก entry data
      properties:
        totalSales: { type: number, description: "salesFromPage + crmSales + hotSales" }
        totalSalesAll: { type: number }
        hotSales: { type: number, description: "Pure revenue (ไม่มีต้นทุน)" }
        totalSpend: { type: number }
        revenue: { type: number }
        profitPage: { type: number, description: "salesFromPage - adminCommission - adCost - (productCost * qty) - packingCost - shippingCost" }
        profitCRM: { type: number, description: "crmSales - (productCost * crmQty)" }
        profitTotal: { type: number, description: "profitPage + profitCRM + profitHot" }
        adPercent: { type: number, nullable: true }
        closeRate: { type: number, nullable: true }
        costPerClick: { type: number, nullable: true }
        roas: { type: number, nullable: true }
        aovPage: { type: number, nullable: true }
        aovCRM: { type: number, nullable: true }
        aovTotal: { type: number, nullable: true }
    Entry:
      type: object
      properties:
        id: { type: integer }
        date: { type: string, format: date, example: "2026-04-07" }
        account:
          $ref: '#/components/schemas/Account'
        product:
          $ref: '#/components/schemas/Product'
        page:
          oneOf:
            - $ref: '#/components/schemas/Page'
            - type: "null"
        crmProduct:
          oneOf:
            - $ref: '#/components/schemas/Product'
            - type: "null"
        entryProducts:
          type: array
          items:
            $ref: '#/components/schemas/EntryProduct'
        adCost: { type: number }
        messages: { type: integer }
        closed: { type: integer }
        orders: { type: integer }
        salesFromPage: { type: number }
        quantity: { type: integer }
        hotSales: { type: number }
        crmOrders: { type: integer }
        crmSales: { type: number }
        crmQty: { type: integer }
        shippingCost: { type: number }
        packingCost: { type: number }
        adminCommission: { type: number }
        note: { type: string, nullable: true }
        calculated:
          $ref: '#/components/schemas/CalculatedFields'

    # --- Request Bodies ---
    EntryInput:
      type: object
      required: [date, accountId, productId, adCost, messages, closed, orders, salesFromPage, quantity]
      properties:
        date: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$', example: "2026-04-07" }
        accountId: { type: integer, minimum: 1 }
        productId: { type: integer, minimum: 1 }
        pageId: { type: integer, minimum: 1, description: "Optional — filter by page" }
        adCost: { type: number, minimum: 0 }
        messages: { type: integer, minimum: 0 }
        closed: { type: integer, minimum: 0 }
        orders: { type: integer, minimum: 0 }
        salesFromPage: { type: number, minimum: 0 }
        quantity: { type: integer, minimum: 0 }
        hotSales: { type: number, minimum: 0, default: 0 }
        crmOrders: { type: integer, minimum: 0, default: 0 }
        crmProductId: { type: integer, minimum: 1 }
        crmSales: { type: number, minimum: 0, default: 0 }
        crmQty: { type: integer, minimum: 0, default: 0 }
        shippingCost: { type: number, minimum: 0, default: 0 }
        packingCost: { type: number, minimum: 0, default: 0 }
        adminCommission: { type: number, minimum: 0, default: 0 }
        products:
          type: array
          description: "Multi-product per entry (junction table)"
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: integer, minimum: 1 }
              quantity: { type: integer, minimum: 0 }
        note: { type: string }
...
    Error:
      type: object
      properties:
        error: { type: string }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
paths:
...
  /api/entries:
    get:
      summary: List entries with filters
      tags: [Entries]
      parameters:
        - name: date
          in: query
          schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        - name: accountId
          in: query
          schema: { type: integer }
        - name: productId
          in: query
          schema: { type: integer }
        - name: pageId
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Paginated entries with calculated fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entry'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create entry (with optional multi-product + stock deduction)
      tags: [Entries]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryInput'
      responses:
        '201':
          description: Created entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...